+++ NVs.OccupancySensor.API/Formatters/RgbImageOutputFormatter.cs
@@ -9,7 +9,7 @@ namespace NVs.OccupancySensor.API.Formatters
 {
     sealed class RgbImageOutputFormatter : OutputFormatter
     {
-        private static readonly Image<Rgb, float> EmptyImage = new Image<Rgb, float>(100, 100);
+        private static readonly Image<Rgb,byte> EmptyImage = new Image<Rgb,byte>(100, 100);
         
         public RgbImageOutputFormatter()
         {
@@ -18,14 +18,14 @@ namespace NVs.OccupancySensor.API.Formatters
 
         protected override bool CanWriteType(Type type)
         {
-            return type == typeof(Image<Rgb, float>);
+            return type == typeof(Image<Rgb,byte>);
         }
 
         public override async Task WriteResponseBodyAsync(OutputFormatterWriteContext context)
         {
             if (context == null) throw new ArgumentNullException(nameof(context));
 
-            var image = context.Object as Image<Rgb, float> ?? EmptyImage;
+            var image = context.Object as Image<Rgb,byte> ?? EmptyImage;
             
             var bytes = image.ToJpegData();
